+++ @navikt/core/react/src/alert/global-alert/root/GlobalAlertRoot.tsx
@@ -24,6 +24,11 @@ interface GlobalAlertProps
    * data-color has no effect on GlobalAlert.
    */
   "data-color"?: never;
+  /**
+   *  Whether title and content are centered or not.
+   * @default true
+   */
+  centered?: boolean;
 }
 
 interface GlobalAlertComponent
@@ -102,14 +107,15 @@ interface GlobalAlertComponent
  * ```
  */
 export const GlobalAlert = forwardRef<HTMLDivElement, GlobalAlertProps>(
-  (props: GlobalAlertProps, forwardedRef) => {
+  ({ centered = true, ...rest }: GlobalAlertProps, forwardedRef) => {
     return (
       <BaseAlert.Root
         ref={forwardedRef}
         role="alert"
-        {...props}
+        {...rest}
         type="strong"
         global
+        data-centered={centered}
       />
     );
   },