+++ @navikt/core/react/src/layout/box/Box.tsx
@@ -123,7 +123,7 @@ export const BoxComponent: OverridableComponent<BoxProps, HTMLDivElement> =
         style: _style,
         asChild,
         ...rest
-      },
+      }: BoxProps & { as?: React.ElementType },
       ref,
     ) => {
       const { cn } = useRenameCSS();