+++ @navikt/core/tokens/src/tokens/colors/create-alpha.ts
@@ -1,4 +1,4 @@
-import Color from "colorjs.io";
+import Color, { type Coords } from "colorjs.io";
 import type { GlobalColorScale } from "../../../internal-types";
 import { AkselColorRole, AkselColorTheme } from "../../../types";
 import { GlobalColorEntry } from "../../tokens.util";
@@ -55,14 +55,30 @@ export function globalConfigWithAlphaTokens({
 function createAlphaColor(targetColor: string, theme: AkselColorTheme) {
   const backgroundColor = semanticRootTokens(theme).bg.default.value;
 
+  const targetCoords = new Color(targetColor).to("srgb").coords;
+  const backgroundCoords = new Color(backgroundColor).to("srgb").coords;
+
   const [r, g, b, a] = getAlphaColor(
-    new Color(targetColor).to("srgb").coords,
-    new Color(backgroundColor).to("srgb").coords,
+    parseAndValidateCoords(targetCoords),
+    parseAndValidateCoords(backgroundCoords),
     255,
     255,
   );
 
   return formatHex(new Color("srgb", [r, g, b], a).toString({ format: "hex" }));
+}
+
+function parseAndValidateCoords(coords: Coords): number[] {
+  const parsedCoords: number[] = [];
+
+  for (const coord of coords) {
+    if (coord === null) {
+      throw new Error(`Color coordinate is undefined: ${coord}`);
+    }
+    parsedCoords.push(coord);
+  }
+
+  return parsedCoords;
 }
 
 // target = background * (1 - alpha) + foreground * alpha