+++ aksel.nav.no/website/app/(routes)/(designsystemet)/grunnleggende/endringslogg/_ui/LogEntry.tsx
@@ -142,9 +142,6 @@ export default function LogEntry({
                     <Link
                       as={NextLink}
                       href={`./endringslogg/${slug}`}
-                      className={
-                        fremhevet ? styles.innholdFremhevet : undefined
-                      }
                       data-color="neutral"
                     >
                       {heading}
@@ -158,11 +155,8 @@ export default function LogEntry({
                     <Hero herobilde={herobilde} />
                   )}
                   <CustomPortableText
-                    className={cl(
-                      fremhevet ? styles.innholdFremhevet : undefined,
-                      styles.portableTextFirstHeading,
-                    )}
-                    data-color={fremhevet ? "aksel-brand-pink" : "brand-blue"}
+                    className={cl(styles.portableTextFirstHeading)}
+                    data-color={fremhevet ? "aksel-brand-pink" : "accent"}
                     value={content as PortableTextBlock[]}
                   />
                 </ShowMore.Content>
@@ -181,7 +175,6 @@ export default function LogEntry({
                   <Link
                     as={NextLink}
                     href={`./endringslogg/${slug}`}
-                    className={fremhevet ? styles.innholdFremhevet : undefined}
                     data-color="neutral"
                   >
                     {heading}
@@ -191,11 +184,8 @@ export default function LogEntry({
                   <Hero herobilde={herobilde} />
                 )}
                 <CustomPortableText
-                  className={cl(
-                    fremhevet ? styles.innholdFremhevet : undefined,
-                    styles.portableTextFirstHeading,
-                  )}
-                  data-color={fremhevet ? "aksel-brand-pink" : undefined}
+                  data-color={fremhevet ? "aksel-brand-pink" : "accent"}
+                  className={cl(styles.portableTextFirstHeading)}
                   value={content as PortableTextBlock[]}
                 />
               </>