+++ aksel.nav.no/website/e2e/sandbox.e2e.ts
@@ -9,7 +9,10 @@ test.describe("Check website sandbox", () => {
     await page.goto(previewUrl);
     await page.waitForLoadState("domcontentloaded");
 
-    const count = await page.locator("#sandbox-wrapper").count();
-    expect(count).toBeGreaterThan(0);
+    const frameLocator = page.frameLocator('iframe[src*="sandbox"]');
+    const sandboxWrapper = frameLocator.locator("#sandbox-wrapper");
+
+    await expect(sandboxWrapper).toBeVisible({ timeout: 2000 });
+    await expect(sandboxWrapper).not.toBeEmpty();
   });
 });