+++ aksel.nav.no/website/pages/eksempler/monthpicker/validation.tsx
@@ -2,9 +2,13 @@ import { useState } from "react";
 import { Box, MonthPicker, useMonthpicker } from "@navikt/ds-react";
 import { withDsExample } from "@/web/examples/withDsExample";
 
+const year = new Date().getFullYear();
+
 const Example = () => {
   const [hasError, setHasError] = useState(false);
   const { monthpickerProps, inputProps } = useMonthpicker({
+    fromDate: new Date(`Aug 23 ${year - 1}`),
+    toDate: new Date(`Aug 23 ${year + 1}`),
     onValidate: (val) => {
       setHasError(!val.isValidMonth);
       console.info(val);
@@ -18,6 +22,7 @@ const Example = () => {
           {...inputProps}
           label="Velg måned"
           error={hasError && "Du må velge måned"}
+          description="Format: mm.yyyy"
         />
       </MonthPicker>
     </Box>